+++ src/main/java/Dao/DeadlineDao.java
@@ -1,4 +1,4 @@
-package Managers;
+package Dao;
 
 import Models.Deadline;
  import Utils.StringUtils;
