+++ src/main/java/Dao/TodoDao.java
@@ -1,4 +1,4 @@
-package Managers;
+package Dao;
 
 import Models.Todo;
 import Utils.StringUtils;
@@ -9,4 +9,6 @@ public class TodoDao {
     String value = StringUtils.getValueOfCommand(input, TodoDao.name, null);
     return new Todo(value);
   }
+
+//  public void add()
 }
